Metric is now 32 bit signed int instead of 16 bit Fixes issue where very lucky users would have metrics clip at 32767%

+++ src/ui.rs
@@ -305,7 +305,7 @@ impl Application for GroupScraper {
             PREMIUM499
         };
         let best_metric =
-            (((robux_per_second / closest_premium.robux_per_second()) - 1.) * 100.) as i16;
+            (((robux_per_second / closest_premium.robux_per_second()) - 1.) * 100.) as i32;
         let robux_count = widget::Text::new(format!(
             "Total robux found: {}\n{} groups checked\n{}% better than {} premium",
             robux_found, self.groups_checked, best_metric, closest_premium.price,
